app_tier/requirements.txt
python-dotenv
boto3
ec2-metadata